requirements.txt
openai==1.12.0
langchain>=0.1.5
langchain-community>=0.0.20
langchain-core>=0.1.5
langchain-google-genai>=0.0.5
streamlit
python-dotenv